refactor: move CLI into a folder
